Task 1 Age Calculator Using "HTML" , "CSS" & "JS"
In this task, I've developed the age calculator using "HTML", "CSS" & "JS" The user will enter its day, month, and year it will calculate its age. I've also applied the validation the day should be from 1 to 31. The month should be from 1 to 12 and the year should be from 1900 to the current year